feat: creating a movie model and the movie api interface
Depdencies added: square retrofit2 and gson. The movie model uses gson
annotations to serialize the movie response keys. The api interface is
created using retrofit with a gson converter. It is currently exposed as
a public api object, later we can move this to more of a repository
pattern so the view model doesn't directly access this api interface.