feat(movie-discovery): simplifying the movie list adapter
Introduces databinding and diffcallback to the adapter
The movie list adapter now uses MovieListItemBinding class generated
from the databinding library to simplify how the layouts are inflated
and how the view holder binds each of the text view values.
The diffcallback companion object now is implemented for the adapter
which removes the not-so-performant notifyDataSetChanged method.