feat(movie-discovery): Refactoring code to use repository pattern
Instead of using the api service directly in the viewmodel, abstracts
the api service into a movie repository. In the future this repository
can use other modules to retrieve the data. i.e. local storage module in
addition to the network module api service that exists now. The goal of
this implementation is to have a single source for the data the
viewmodels will need. This makes it easier to reason about the code as
every part of the app is following the single reposibility rule. In
addition, makes it easier to write unit tests.
This is implemented with the help of dependency injection using
android's recommended dagger hilt:
https://developer.android.com/training/dependency-injection,